requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==6.0.2
psycopg2==2.7.4 --no-binary psycopg2
python-dateutil==2.7.3
PyJWT==1.6.4
isort==4.3.4

# Django
django==2.2.9

# Django REST Framework
djangorestframework==3.9.3
django-filter==2.1.0
djangorestframework-simplejwt==4.1.4

# Cors
django-cors-headers==2.3.0


# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==18.3.0

# Static files
whitenoise==4.1.2

# Celery
redis==3.2.1
django-redis==4.10.0
celery==4.2.1
flower==0.9.2
tornado>=4.2.0,<6.0.0